[FIX]: API url 이름 변경 및 컨트롤러, 서비스 이름 변경, isFull 의존 삭제

+++ src/main/java/umc/parasol/domain/history/domain/History.java
@@ -46,7 +46,7 @@ public class History extends BaseEntity {
     private Umbrella umbrella;
 
     @NotNull(message = "가격이 설정되어 있어야 합니다.")
-    private Long cost;
+    private int cost;
     
     @Enumerated(EnumType.STRING)
     @NotNull(message = "절차가 설정되어 있어야 합니다.")
